# PERF: merge on sorted MultiIndex